requirements.txt
asgiref==3.7.2
Django==4.1.9
django-taggit==3.0.0
Markdown==3.4.1
psycopg2-binary==2.9.3
python-dotenv==1.0.0
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
